nscms/requirements.txt
Django
FeinCMS
django-photologue
django-taggit
django-extensions
django-ckeditor
South
django-mptt
